moved conversion function
