+++ libs/langchain-google-genai/package.json
@@ -34,10 +34,8 @@
   "author": "LangChain",
   "license": "MIT",
   "dependencies": {
-    "@google-ai/generativelanguage": "^2.0.0",
     "@google/generative-ai": "^0.1.0",
-    "@langchain/core": "~0.1.1",
-    "google-auth-library": "^8.9.0"
+    "@langchain/core": "~0.1.1"
   },
   "devDependencies": {
     "@jest/globals": "^29.5.0",
